* Check file size of download if possible

+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/BackgroundJob.kt
@@ -224,6 +224,15 @@ class BackgroundJob(context: Context, workerParams: WorkerParameters) :
                 app.downloadedFileCache.deleteZipFile(context)
             }
 
+            if (latestUpdate.fileSizeBytes != null && latestUpdate.fileSizeBytes != downloadFile.length()) {
+                val expected = latestUpdate.fileSizeBytes
+                val actual = downloadFile.length()
+                throw NetworkException(
+                    "Wrong file was downloaded. It should be $expected bytes long but " +
+                            "actual it was $actual bytes. FFUpdater will retry the download later."
+                )
+            }
+
             notificationRemover.removeDownloadRunningNotification(context, app)
             app.downloadedFileCache.deleteAllExceptLatestApkFile(context, latestUpdate)
             true
